Use icons and text for player info
