+++ src/Player.css
@@ -6,7 +6,7 @@
 .general {
     font-size: 3vw;
     color: white;
-    margin-left: 1vw;
+    margin-left: 2vw;
 }
 
 .note {
